backend/requirements.txt
# ── Core ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Database ──
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# ── Redis ──
redis[hiredis]==5.2.1
rq==2.1.0

# ── LLM Providers ──
openai==1.58.1
google-generativeai==0.8.4

# ── Speech ──
deepgram-sdk==3.8.0
google-cloud-texttospeech==2.21.1

# ── Utilities ──
httpx==0.28.1
orjson==3.10.13

# ── Testing ──
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
